+++ index.html
@@ -23,7 +23,7 @@
     </section>
 
     <!-- Контейнер для изображений от других пользователей -->
-    <section class="pictures  container">
+    <section class="pictures  container js-pictures">
       <h2 class="pictures__title  visually-hidden">Фотографии других пользователей</h2>
 
       <!-- Поле для загрузки нового изображения на сайт -->
@@ -232,7 +232,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="main.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
