+++ js/main.js
@@ -1,4 +1,4 @@
-import {getPhotosDescription} from './data.js';
+import { getPhotosDescription } from './mocks.js';
+import { getThumbnails } from './thumbnails.js';
 
-// eslint-disable-next-line no-unused-expressions
-getPhotosDescription();
+getThumbnails(getPhotosDescription);
